requirements.txt
astroid==1.5.3
autopep8==1.3.2
certifi==2017.11.5
chardet==3.0.4
click==6.7
Django==1.11.3
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.0.6
pycodestyle==2.3.1
pylint==1.7.2
python-dotenv==0.7.1
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
twython==3.6.0
urllib3==1.22
virtualenv==15.1.0
wrapt==1.10.10
